fix: CoursesCards component
