--- client/src/components/CourseCard/courseCard.scss
@@ -1,41 +1,0 @@
-@import '../../styles/';
-
-.course-card {
-  background-color: $primary-white;
-  border-radius: 0.75rem;
-  box-shadow: rgba(0, 0, 0, 0.19) 0px 10px 20px, rgba(0, 0, 0, 0.23) 0px 6px 6px;
-  padding: 1.5rem 1rem;
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-  justify-content: space-between;
-  @media (max-width: 45rem) {
-    width: 100%;
-  }
-  @media (min-width: 45rem) and (max-width: 80rem) {
-    width: calc(50% - 1rem);
-  }
-  @media (min-width: 80rem) {
-    width: calc(33.333% - 1.333rem);
-  }
-  .intro {
-    img {
-      width: 100%;
-      height: 15rem;
-      object-fit: cover;
-    }
-
-    h2 {
-      padding: 0 0 0.75rem 0;
-    }
-  }
-  .see-more-btn {
-    box-shadow: none;
-    background-color: $background-gray;
-    border-radius: 0.25rem;
-    &:hover {
-      background-color: $primary-green;
-      color: white;
-    }
-  }
-}
