+++ client/src/pages/AllCourses/index.jsx
@@ -3,31 +3,31 @@ import CourseService from 'services/course.service';
 // Font Awesome
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faMagnifyingGlass } from '@fortawesome/free-solid-svg-icons';
-import CourseCard from 'components/CourseCard';
+import CourseCards from 'components/CourseCards';
 import './allCourses.scss';
 
 const AllCourses = ({ currentUser, setIsModalOpen }) => {
   // state 用來儲存從API當中所獲得的Course Data
-  const [courseData, setCourseData] = useState([]);
+  const [allCoursesData, setAllCoursesData] = useState([]);
   const [searchInput, setSearchInput] = useState('');
-  const [searchResult, setSearchResult] = useState([]);
+  const [searchCourseData, setSearchCourseData] = useState([]);
   const [showMessage, setShowMessage] = useState(false);
   const [showNotFoundMsg, setShowNotFoundMsg] = useState(false);
 
   const handleChangeInput = (e) => {
     setSearchInput(e.target.value);
   };
+
   const handleSearch = () => {
-    setSearchResult([]); // 清空搜尋結果
+    setSearchCourseData([]); // 清空搜尋結果
     setShowMessage(false); // 清空提示
     setShowNotFoundMsg(false); // 清空查無資料提示
-
     if (searchInput.trim() !== '') {
       CourseService.getCourseByName(searchInput.trim())
         .then((res) => {
           console.log({ res });
           if (res.data.length !== 0) {
-            setSearchResult(res.data); // 取得資料
+            setSearchCourseData(res.data); // 取得資料
           } else {
             setShowNotFoundMsg(true); // 打開查無資料提示
           }
@@ -38,7 +38,6 @@ const AllCourses = ({ currentUser, setIsModalOpen }) => {
     } else {
       setShowMessage(true); // 打開提示
     }
-
     setSearchInput(''); // 清空搜尋欄位的值
   };
 
@@ -46,9 +45,9 @@ const AllCourses = ({ currentUser, setIsModalOpen }) => {
   useEffect(() => {
     // console.log('Using effect!');
     CourseService.getAllCourses()
-      .then((data) => {
-        console.log({ data });
-        setCourseData(data.data); // Array
+      .then((res) => {
+        console.log({ res });
+        setAllCoursesData(res.data); // Array
       })
       .catch((error) => {
         console.log(error);
@@ -73,18 +72,11 @@ const AllCourses = ({ currentUser, setIsModalOpen }) => {
             </div>
           )}
           {showMessage && <div className='message'>Please enter the course name you want to search.</div>}
-          {searchResult &&
-            searchResult.map((course) => (
-              <CourseCard key={course._id} course={course} currentUser={currentUser} setIsModalOpen={setIsModalOpen}></CourseCard>
-            ))}
+          {searchCourseData && <CourseCards courses={searchCourseData} currentUser={currentUser} setIsModalOpen={setIsModalOpen}></CourseCards>}
         </div>
-        {searchResult && searchResult.length > 0 && <div className='dividing-line'></div>}
+        {searchCourseData && searchCourseData.length > 0 && <div className='dividing-line'></div>}
 
-        <div className='all-courses'>
-          {courseData.map((course) => {
-            return <CourseCard key={course._id} course={course} currentUser={currentUser} setIsModalOpen={setIsModalOpen}></CourseCard>;
-          })}
-        </div>
+        <CourseCards courses={allCoursesData} currentUser={currentUser} setIsModalOpen={setIsModalOpen}></CourseCards>
       </section>
     </main>
   );
